Visiteur nyeja

+++ backend/src/main/java/com/projetCloud/projetCloud/dto/RecapSignalementDto.java
@@ -1,12 +1,14 @@
 package com.projetCloud.projetCloud.dto;
 
+import java.math.BigDecimal;
+
 public class RecapSignalementDto {
     private Long nombrePoints;
-    private Double totalSurface;
-    private Double totalBudget;
-    private Double avancementPourcent;
+    private BigDecimal totalSurface;
+    private BigDecimal totalBudget;
+    private BigDecimal avancementPourcent;
 
-    public RecapSignalementDto(Long nombrePoints, Double totalSurface, Double totalBudget, Double avancementPourcent) {
+    public RecapSignalementDto(Long nombrePoints, BigDecimal totalSurface, BigDecimal totalBudget, BigDecimal avancementPourcent) {
         this.nombrePoints = nombrePoints;
         this.totalSurface = totalSurface;
         this.totalBudget = totalBudget;
@@ -16,10 +18,10 @@ public class RecapSignalementDto {
     // Getters et setters
     public Long getNombrePoints() { return nombrePoints; }
     public void setNombrePoints(Long nombrePoints) { this.nombrePoints = nombrePoints; }
-    public Double getTotalSurface() { return totalSurface; }
-    public void setTotalSurface(Double totalSurface) { this.totalSurface = totalSurface; }
-    public Double getTotalBudget() { return totalBudget; }
-    public void setTotalBudget(Double totalBudget) { this.totalBudget = totalBudget; }
-    public Double getAvancementPourcent() { return avancementPourcent; }
-    public void setAvancementPourcent(Double avancementPourcent) { this.avancementPourcent = avancementPourcent; }
+    public BigDecimal getTotalSurface() { return totalSurface; }
+    public void setTotalSurface(BigDecimal totalSurface) { this.totalSurface = totalSurface; }
+    public BigDecimal getTotalBudget() { return totalBudget; }
+    public void setTotalBudget(BigDecimal totalBudget) { this.totalBudget = totalBudget; }
+    public BigDecimal getAvancementPourcent() { return avancementPourcent; }
+    public void setAvancementPourcent(BigDecimal avancementPourcent) { this.avancementPourcent = avancementPourcent; }
 }
